refactor: extract ChatPanel from ChatPage
Move the message-list, composer, SSE streaming, consent modal, quota
line, and attachment-chip logic from ChatPage into a standalone
ChatPanel component at web/src/components/ChatPanel/.

ChatPanel accepts initialPrompt, initialMessages, and initialConversationId
to seed its state on first render. ChatPage now uses the key pattern to
remount ChatPanel when the user selects a different conversation, keeping
all message state inside the panel while the page manages the sidebar.

Four Vitest tests cover: renders with initialPrompt, calls
/api/chat/message, handles consent_required, handles rate_limit.
