+++ web/src/pages/Chat/ChatPage.tsx
@@ -1,46 +1,11 @@
-import { useEffect, useRef, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { useUserLocals } from '../../lib/hooks/useUserLocals';
-import { get, post, postMultipart, patch, del } from '../../lib/backend/api';
-import ConsentModal from '../../components/ConsentModal/ConsentModal';
-import SendIcon from '../../components/icons/SendIcon';
-import AssistantMarkdown from './AssistantMarkdown';
-import CardPreview from './CardPreview';
+import { get, patch, del } from '../../lib/backend/api';
+import ChatPanel, { type Message } from '../../components/ChatPanel/ChatPanel';
 import ConversationsSidebar, {
   ConversationSummary,
 } from './ConversationsSidebar';
 import styles from './ChatPage.module.css';
-
-interface ChatCard {
-  front: string;
-  back: string;
-}
-
-interface Message {
-  role: 'user' | 'assistant';
-  content: string;
-  contentBefore?: string;
-  contentAfter?: string;
-  cards?: ChatCard[];
-}
-
-interface ApiDonePayload {
-  content: string;
-  conversationId: number;
-  contentBefore?: string;
-  contentAfter?: string;
-  cards?: ChatCard[];
-}
-
-interface ApiErrorPayload {
-  type: 'rate_limit' | 'server_error' | 'conversation_not_found' | 'consent_required';
-  resetDate?: string;
-}
-
-interface ApiUsageResponse {
-  used: number;
-  limit: number | null;
-}
 
 interface ApiConversationsResponse {
   conversations: ConversationSummary[];
@@ -51,7 +16,7 @@ interface ApiConversationDetailMessage {
   role: 'user' | 'assistant';
   content: string;
   createdAt: string;
-  cards?: ChatCard[];
+  cards?: { front: string; back: string }[];
   contentBefore?: string;
   contentAfter?: string;
 }
@@ -65,93 +30,20 @@ interface ApiConversationDetailResponse {
   messages: ApiConversationDetailMessage[];
 }
 
-const DRAFT_DEBOUNCE_MS = 500;
-
-type ChipState = 'idle' | 'uploading' | 'failed';
-
-interface AttachmentChip {
-  id: string;
-  file: File;
-  state: ChipState;
-  retryCount: number;
-}
-
-const ALLOWED_TYPES = new Set([
-  'application/pdf',
-  'image/png',
-  'image/jpeg',
-  'image/gif',
-  'image/webp',
-]);
-
-const MAX_FILE_BYTES = 10 * 1024 * 1024;
-const MAX_TOTAL_BYTES = 25 * 1024 * 1024;
-const MAX_FILE_COUNT = 5;
-
 const STARTER_PROMPTS = [
   "Make 10 cards from notes I'll paste",
   'Explain a concept, then make cards',
   'Turn this into cloze cards: [paste]',
 ];
 
-const FREE_MONTHLY_LIMIT = 20;
-
-function formatResetDate(iso: string): string {
-  try {
-    return new Date(iso).toLocaleDateString(undefined, { month: 'long', day: 'numeric' });
-  } catch {
-    return iso;
-  }
-}
-
-function visibleStreamingText(text: string): string {
-  const fenceIndex = text.search(/(?:^|\n)```json/);
-  if (fenceIndex !== -1) return text.slice(0, fenceIndex);
-  const rawArrayIndex = text.search(/(?:^|\n)\s*\[\s*\{/);
-  return rawArrayIndex === -1 ? text : text.slice(0, rawArrayIndex);
-}
-
-async function downloadDeck(cards: ChatCard[], deckName: string): Promise<void> {
-  const response = await post('/api/chat/deck', { cards, deckName });
-  if (!response.ok) {
-    throw new Error('Failed to generate deck');
-  }
-  const blob = await response.blob();
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = `${deckName}.apkg`;
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-function formatFileSize(bytes: number): string {
-  if (bytes < 1024 * 1024) {
-    return `${Math.round(bytes / 1024)} KB`;
-  }
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-function chipIcon(mimeType: string): string {
-  return mimeType === 'application/pdf' ? '📄' : '🖼';
-}
-
-function truncateName(name: string, max: number): string {
-  if (name.length <= max) return name;
-  const ext = name.lastIndexOf('.');
-  if (ext > 0 && name.length - ext <= 6) {
-    const truncated = name.slice(0, max - 3 - (name.length - ext));
-    return `${truncated}…${name.slice(ext)}`;
-  }
-  return `${name.slice(0, max - 1)}…`;
+interface PanelSeed {
+  key: string;
+  conversationId: number | null;
+  messages: Message[];
+  draft: string;
 }
 
 export default function ChatPage() {
-  const { data: userLocals, refetch: refetchUserLocals } = useUserLocals();
-  const isPatreon = userLocals?.user?.patreon === true;
-  const hasConsented = userLocals?.user?.chat_consent_at != null;
-  const [showConsentModal, setShowConsentModal] = useState(false);
-
   const [searchParams] = useSearchParams();
   const cameFromUpload = searchParams.get('from') === 'upload';
   const uploadFilename = searchParams.get('filename');
@@ -161,37 +53,14 @@ export default function ChatPage() {
 
   const [conversations, setConversations] = useState<ConversationSummary[]>([]);
   const [activeConversationId, setActiveConversationId] = useState<number | null>(null);
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [expandedUserMessages, setExpandedUserMessages] = useState<Set<number>>(new Set());
-  const [streamingText, setStreamingText] = useState('');
-  const [inputValue, setInputValue] = useState(prefilledPrompt);
-  const [isLoading, setIsLoading] = useState(false);
-  const [networkError, setNetworkError] = useState<string | null>(null);
-  const [limitReached, setLimitReached] = useState(false);
-  const [resetDate, setResetDate] = useState<string | null>(null);
-  const [messagesUsedThisMonth, setMessagesUsedThisMonth] = useState(0);
-  const [chips, setChips] = useState<AttachmentChip[]>([]);
-  const [isDragging, setIsDragging] = useState(false);
-
-  const bottomRef = useRef<HTMLDivElement>(null);
-  const messageListRef = useRef<HTMLDivElement>(null);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
-  const lastSavedDraftRef = useRef<string>('');
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    get('/api/chat/usage', { redirect: false })
-      .then((data: ApiUsageResponse | undefined) => {
-        if (data != null) {
-          setMessagesUsedThisMonth(data.used);
-          if (data.limit != null && data.used >= data.limit) {
-            setLimitReached(true);
-          }
-        }
-      })
-      .catch(() => {
-      });
-  }, []);
+  const [panelSeed, setPanelSeed] = useState<PanelSeed>({
+    key: 'new',
+    conversationId: null,
+    messages: [],
+    draft: prefilledPrompt,
+  });
+  const [loadError, setLoadError] = useState<string | null>(null);
+  const [showEmptyState, setShowEmptyState] = useState(true);
 
   useEffect(() => {
     get('/api/chat/conversations', { redirect: false })
@@ -203,91 +72,6 @@ export default function ChatPage() {
       .catch(() => {
       });
   }, []);
-
-  useEffect(() => {
-    const el = messageListRef.current;
-    if (el == null) return;
-    const distanceFromBottom = el.scrollHeight - el.scrollTop - el.clientHeight;
-    if (distanceFromBottom < 120) {
-      bottomRef.current?.scrollIntoView({ behavior: streamingText.length > 0 ? 'auto' : 'smooth' });
-    }
-  }, [messages, isLoading, streamingText]);
-
-  useEffect(() => {
-    if (activeConversationId == null) return;
-    if (inputValue === lastSavedDraftRef.current) return;
-    const conversationId = activeConversationId;
-    const draft = inputValue;
-    const handle = setTimeout(() => {
-      patch(`/api/chat/conversations/${conversationId}/draft`, {
-        content: draft.length === 0 ? null : draft,
-      })
-        .then(() => {
-          lastSavedDraftRef.current = draft;
-        })
-        .catch(() => {
-        });
-    }, DRAFT_DEBOUNCE_MS);
-    return () => clearTimeout(handle);
-  }, [inputValue, activeConversationId]);
-
-  const remainingMessages = FREE_MONTHLY_LIMIT - messagesUsedThisMonth;
-
-  const readyChips = chips.filter((c) => c.state === 'idle');
-  const canSend = (inputValue.trim().length > 0 || readyChips.length > 0) && !isLoading && !limitReached;
-
-  function addFiles(files: File[]) {
-    setNetworkError(null);
-
-    const disallowed = files.filter((f) => !ALLOWED_TYPES.has(f.type));
-    if (disallowed.length > 0) {
-      if (disallowed.length === 1) {
-        setNetworkError(`Can't attach ${disallowed[0].name}. Only PDF and image files work here.`);
-      } else {
-        setNetworkError(`Can't attach ${disallowed.length} files. Only PDF and image files work here.`);
-      }
-      return;
-    }
-
-    const oversized = files.find((f) => f.size > MAX_FILE_BYTES);
-    if (oversized != null) {
-      setNetworkError(`${oversized.name} is ${formatFileSize(oversized.size)}. The per-file limit is 10 MB.`);
-      return;
-    }
-
-    const currentTotal = chips.reduce((s, c) => s + c.file.size, 0);
-    const newTotal = files.reduce((s, f) => s + f.size, currentTotal);
-    if (newTotal > MAX_TOTAL_BYTES) {
-      setNetworkError(`That's ${formatFileSize(newTotal)} total. A message can carry up to 25 MB across all files.`);
-      return;
-    }
-
-    const currentCount = chips.length;
-    const allowedCount = Math.max(0, MAX_FILE_COUNT - currentCount);
-    const toAdd = files.slice(0, allowedCount);
-
-    setChips((prev) => [
-      ...prev,
-      ...toAdd.map<AttachmentChip>((f) => ({
-        id: crypto.randomUUID(),
-        file: f,
-        state: 'idle',
-        retryCount: 0,
-      })),
-    ]);
-  }
-
-  function removeChip(id: string) {
-    setChips((prev) => prev.filter((c) => c.id !== id));
-  }
-
-  function retryChip(id: string) {
-    setChips((prev) =>
-      prev.map((c) =>
-        c.id === id ? { ...c, state: 'idle', retryCount: c.retryCount + 1 } : c
-      )
-    );
-  }
 
   function upsertConversation(id: number, title: string): void {
     setConversations((prev) => {
@@ -305,47 +89,37 @@ export default function ChatPage() {
 
   async function handleSelectConversation(id: number) {
     if (id === activeConversationId) return;
+    setLoadError(null);
     setActiveConversationId(id);
-    setMessages([]);
-    setExpandedUserMessages(new Set());
-    setStreamingText('');
-    setNetworkError(null);
-    setInputValue('');
-    lastSavedDraftRef.current = '';
-    setChips([]);
     try {
       const data = (await get(`/api/chat/conversations/${id}`, { redirect: false })) as
         | ApiConversationDetailResponse
         | undefined;
       if (data == null) return;
-      setMessages(
-        data.messages.map((m) => ({
-          role: m.role,
-          content: m.content,
-          ...(m.cards == null ? {} : { cards: m.cards }),
-          ...(m.contentBefore == null ? {} : { contentBefore: m.contentBefore }),
-          ...(m.contentAfter == null ? {} : { contentAfter: m.contentAfter }),
-        }))
-      );
-      if (data.draft != null && data.draft.length > 0) {
-        setInputValue(data.draft);
-        lastSavedDraftRef.current = data.draft;
-      }
+      const loaded: Message[] = data.messages.map((m) => ({
+        role: m.role,
+        content: m.content,
+        ...(m.cards == null ? {} : { cards: m.cards }),
+        ...(m.contentBefore == null ? {} : { contentBefore: m.contentBefore }),
+        ...(m.contentAfter == null ? {} : { contentAfter: m.contentAfter }),
+      }));
+      setPanelSeed({
+        key: `conv-${id}`,
+        conversationId: id,
+        messages: loaded,
+        draft: data.draft ?? '',
+      });
+      setShowEmptyState(loaded.length === 0);
     } catch {
-      setNetworkError("Couldn't load this conversation.");
+      setLoadError("Couldn't load this conversation.");
     }
   }
 
   function handleNewConversation() {
     setActiveConversationId(null);
-    setMessages([]);
-    setExpandedUserMessages(new Set());
-    setStreamingText('');
-    setNetworkError(null);
-    setInputValue('');
-    lastSavedDraftRef.current = '';
-    setChips([]);
-    textareaRef.current?.focus();
+    setPanelSeed({ key: `new-${Date.now()}`, conversationId: null, messages: [], draft: '' });
+    setShowEmptyState(true);
+    setLoadError(null);
   }
 
   async function handleRenameConversation(id: number, title: string) {
@@ -357,11 +131,11 @@ export default function ChatPage() {
       const response = await patch(`/api/chat/conversations/${id}`, { title });
       if (!response.ok) {
         setConversations(previous);
-        setNetworkError("Couldn't rename this conversation.");
+        setLoadError("Couldn't rename this conversation.");
       }
     } catch {
       setConversations(previous);
-      setNetworkError("Couldn't rename this conversation.");
+      setLoadError("Couldn't rename this conversation.");
     }
   }
 
@@ -376,208 +150,15 @@ export default function ChatPage() {
       if (response == null) return;
       if (!response.ok) {
         setConversations(previous);
-        setNetworkError("Couldn't delete this conversation.");
+        setLoadError("Couldn't delete this conversation.");
       }
     } catch {
       setConversations(previous);
-      setNetworkError("Couldn't delete this conversation.");
+      setLoadError("Couldn't delete this conversation.");
     }
   }
-
-  async function sendMessage(content: string) {
-    if (!content.trim() && readyChips.length === 0) return;
-
-    const userMessage: Message = { role: 'user', content };
-    const nextMessages = [...messages, userMessage];
-    setMessages(nextMessages);
-    setInputValue('');
-    setChips([]);
-    setNetworkError(null);
-    setIsLoading(true);
-    setStreamingText('');
-
-    const history = nextMessages.slice(-10).map((m) => ({ role: m.role, content: m.content }));
-
-    let response: Response;
-    if (readyChips.length > 0) {
-      const formData = new FormData();
-      formData.append('content', content);
-      formData.append('history', JSON.stringify(history));
-      if (activeConversationId != null) {
-        formData.append('conversationId', String(activeConversationId));
-      }
-      for (const chip of readyChips) {
-        formData.append('files', chip.file, chip.file.name);
-      }
-      try {
-        response = await postMultipart('/api/chat/message', formData);
-      } catch {
-        setNetworkError("Couldn't send this message. Try again.");
-        setIsLoading(false);
-        return;
-      }
-    } else {
-      try {
-        response = await post('/api/chat/message', {
-          content,
-          history,
-          conversationId: activeConversationId,
-        });
-      } catch {
-        setNetworkError("Couldn't send this message. Try again.");
-        setIsLoading(false);
-        return;
-      }
-    }
-
-    if (!response.ok) {
-      const data = await response.json().catch(() => ({})) as { error?: string };
-      setNetworkError(data.error ?? "Couldn't send this message. Try again.");
-      setIsLoading(false);
-      return;
-    }
-
-    if (response.body == null) {
-      setNetworkError("Couldn't send this message. Try again.");
-      setIsLoading(false);
-      return;
-    }
-
-    const reader = response.body.getReader();
-    const decoder = new TextDecoder();
-    let buffer = '';
-
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-
-        buffer += decoder.decode(value, { stream: true });
-        const events = buffer.split('\n\n');
-        buffer = events.pop() ?? '';
-
-        for (const rawEvent of events) {
-          if (!rawEvent.trim()) continue;
-
-          const lines = rawEvent.split('\n');
-          let eventType = '';
-          let data = '';
-
-          for (const line of lines) {
-            if (line.startsWith('event: ')) {
-              eventType = line.slice(7).trim();
-            } else if (line.startsWith('data: ')) {
-              data = line.slice(6);
-            }
-          }
-
-          if (eventType === 'token') {
-            const text = JSON.parse(data) as string;
-            setStreamingText((prev) => prev + text);
-          } else if (eventType === 'done') {
-            const result = JSON.parse(data) as ApiDonePayload;
-            setMessages((prev) => [
-              ...prev,
-              {
-                role: 'assistant',
-                content: result.content,
-                contentBefore: result.contentBefore,
-                contentAfter: result.contentAfter,
-                cards: result.cards,
-              },
-            ]);
-            setStreamingText('');
-            setMessagesUsedThisMonth((n) => n + 1);
-            setActiveConversationId(result.conversationId);
-            lastSavedDraftRef.current = '';
-            const provisionalTitle =
-              content.length > 60 ? `${content.slice(0, 60).trimEnd()}…` : content;
-            upsertConversation(result.conversationId, provisionalTitle);
-          } else if (eventType === 'error') {
-            const err = JSON.parse(data) as ApiErrorPayload;
-            if (err.type === 'rate_limit') {
-              setLimitReached(true);
-              if (err.resetDate != null) setResetDate(err.resetDate);
-            } else if (err.type === 'conversation_not_found') {
-              setNetworkError('This conversation is gone. Start a new one.');
-              setActiveConversationId(null);
-            } else if (err.type === 'consent_required') {
-              setShowConsentModal(true);
-            } else {
-              setNetworkError("Couldn't send this message. Try again.");
-            }
-          }
-        }
-      }
-    } catch {
-      setNetworkError("Couldn't send this message. Try again.");
-    } finally {
-      setIsLoading(false);
-      setStreamingText('');
-    }
-  }
-
-  function handleKeyDown(e: React.KeyboardEvent<HTMLTextAreaElement>) {
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      if (canSend) {
-        sendMessage(inputValue);
-      }
-    }
-  }
-
-  function toggleUserMessage(index: number) {
-    setExpandedUserMessages((prev) => {
-      const next = new Set(prev);
-      if (next.has(index)) {
-        next.delete(index);
-      } else {
-        next.add(index);
-      }
-      return next;
-    });
-  }
-
-  function handleSaveAsDeck(cards: ChatCard[], deckName: string) {
-    downloadDeck(cards, deckName).catch(() => {
-      setNetworkError("Couldn't generate the deck. Try again.");
-    });
-  }
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    setIsDragging(true);
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    if (e.currentTarget.contains(e.relatedTarget as Node)) return;
-    setIsDragging(false);
-  }, []);
-
-  const handleDrop = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    setIsDragging(false);
-    const files = Array.from(e.dataTransfer.files);
-    if (files.length > 0) {
-      addFiles(files);
-    }
-  }, [chips]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  const isCardStreaming = /(?:^|\n)```json/.test(streamingText) || /(?:^|\n)\s*\[\s*\{/.test(streamingText);
-
-  const hasMessages = messages.length > 0;
 
   return (
-    <>
-    {(showConsentModal || (!hasConsented && userLocals != null)) && (
-      <ConsentModal
-        onAccept={async () => {
-          await refetchUserLocals();
-          setShowConsentModal(false);
-        }}
-        onDismiss={() => setShowConsentModal(false)}
-      />
-    )}
     <div className={styles.layout} data-hj-suppress>
       <ConversationsSidebar
         conversations={conversations}
@@ -588,78 +169,7 @@ export default function ChatPage() {
         onDelete={handleDeleteConversation}
       />
       <div className={styles.container}>
-        {hasMessages || isLoading ? (
-          <div className={styles.messageList} ref={messageListRef}>
-            <div className={styles.messageListInner}>
-            {messages.map((m, i) => (
-              <div
-                key={i}
-                className={`${styles.message} ${m.role === 'user' ? styles.messageUser : styles.messageAssistant} ${m.role === 'assistant' && m.cards != null && m.cards.length > 0 ? styles.messageAssistantWithCards : ''}`}
-              >
-                {m.role === 'user' && (() => {
-                  const isLong = m.content.length > 600 || m.content.split('\n').length > 12;
-                  const isExpanded = expandedUserMessages.has(i);
-                  return (
-                    <>
-                      <div
-                        className={`${styles.messageBubble} ${styles.messageBubbleUser} ${isLong ? styles.userBubbleCollapsible : ''} ${isLong && !isExpanded ? styles.userBubbleClamped : ''}`}
-                      >
-                        {m.content}
-                      </div>
-                      {isLong && (
-                        <button
-                          type="button"
-                          className={styles.expandToggle}
-                          onClick={() => toggleUserMessage(i)}
-                          aria-expanded={isExpanded}
-                        >
-                          {isExpanded ? 'Show less' : 'Show full message'}
-                        </button>
-                      )}
-                    </>
-                  );
-                })()}
-                {m.role === 'assistant' && (
-                  <>
-                    {m.contentBefore != null && (
-                      <AssistantMarkdown>{m.contentBefore}</AssistantMarkdown>
-                    )}
-                    {m.cards != null && m.cards.length > 0 && (
-                      <CardPreview
-                        cards={m.cards}
-                        onSave={(deckName) => handleSaveAsDeck(m.cards!, deckName)}
-                      />
-                    )}
-                    {m.contentAfter != null && (
-                      <AssistantMarkdown>{m.contentAfter}</AssistantMarkdown>
-                    )}
-                    {m.cards == null && (
-                      <AssistantMarkdown>{m.content}</AssistantMarkdown>
-                    )}
-                  </>
-                )}
-              </div>
-            ))}
-            {isLoading && (
-              <div className={`${styles.message} ${styles.messageAssistant}`}>
-                {streamingText.length > 0 ? (
-                  <>
-                    <AssistantMarkdown isStreaming={!isCardStreaming}>
-                      {visibleStreamingText(streamingText)}
-                    </AssistantMarkdown>
-                    {isCardStreaming && (
-                      <span className={styles.buildingCards}>Building cards</span>
-                    )}
-                  </>
-                ) : (
-                  <div className={styles.messageSkeleton} />
-                )}
-              </div>
-            )}
-            <div ref={bottomRef} />
-            </div>
-          </div>
-        ) : (
+        {showEmptyState && (
           <div className={styles.emptyState}>
             <h1 className={styles.emptyHeading}>Chat</h1>
             <p className={styles.emptySubhead}>
@@ -674,10 +184,9 @@ export default function ChatPage() {
                     key={prompt}
                     type="button"
                     className={styles.starterChip}
-                    onClick={() => {
-                      setInputValue(prompt);
-                      textareaRef.current?.focus();
-                    }}
+                    onClick={() =>
+                      setPanelSeed((prev) => ({ ...prev, draft: prompt }))
+                    }
                   >
                     {prompt}
                   </button>
@@ -686,149 +195,28 @@ export default function ChatPage() {
             )}
           </div>
         )}
-
-        <div className={styles.inputArea}>
-          {limitReached && resetDate != null && (
-            <div className={styles.limitPanel}>
-              <span>
-                You've used all {FREE_MONTHLY_LIMIT} messages this month. Resets {formatResetDate(resetDate)}.
-              </span>
-              <a
-                href="https://www.patreon.com/alemayhu"
-                target="_blank"
-                rel="noreferrer"
-                className={styles.limitPanelLink}
-              >
-                See Patreon plans
-              </a>
-            </div>
-          )}
-          <div
-            role="region"
-            aria-label="Chat composer with file drop zone"
-            className={`${styles.composerCard} ${isDragging ? styles.composerCardDragging : ''}`}
-            onDragOver={handleDragOver}
-            onDragLeave={handleDragLeave}
-            onDrop={handleDrop}
-          >
-            {isDragging && (
-              <div className={styles.dropOverlay}>
-                <span className={styles.dropOverlayTitle}>Drop to attach</span>
-                <span className={styles.dropOverlaySub}>PDF or image, up to 10 MB each</span>
-              </div>
-            )}
-            {chips.length > 0 && (
-              <div className={styles.chipStrip}>
-                {chips.map((chip) => (
-                  <div
-                    key={chip.id}
-                    className={`${styles.chip} ${chip.state === 'failed' ? styles.chipError : ''}`}
-                  >
-                    <span className={styles.chipIcon}>{chipIcon(chip.file.type)}</span>
-                    <span
-                      className={styles.chipName}
-                      title={chip.file.name}
-                    >
-                      {truncateName(chip.file.name, 32)}
-                    </span>
-                    <span className={styles.chipSeparator}> · </span>
-                    {chip.state === 'uploading' && (
-                      <>
-                        <span className={`${styles.chipSize} ${styles.chipSizeError}`}>
-                          <span className={styles.spinnerSmall} />
-                        </span>
-                        <span className={styles.chipSize}>Uploading</span>
-                      </>
-                    )}
-                    {chip.state === 'failed' && (
-                      <>
-                        <span className={`${styles.chipSize} ${styles.chipSizeError}`}>Upload failed</span>
-                        <button
-                          type="button"
-                          className={styles.chipRetry}
-                          onClick={() => retryChip(chip.id)}
-                        >
-                          Retry
-                        </button>
-                      </>
-                    )}
-                    {chip.state === 'idle' && (
-                      <span className={styles.chipSize}>{formatFileSize(chip.file.size)}</span>
-                    )}
-                    <button
-                      type="button"
-                      className={styles.chipRemove}
-                      aria-label={`Remove ${chip.file.name}`}
-                      onClick={() => removeChip(chip.id)}
-                    >
-                      ×
-                    </button>
-                  </div>
-                ))}
-              </div>
-            )}
-            <div className={styles.composerBottom}>
-              <textarea
-                ref={textareaRef}
-                className={styles.textarea}
-                value={inputValue}
-                onChange={(e) => setInputValue(e.target.value)}
-                onKeyDown={handleKeyDown}
-                placeholder="Ask a study question, paste notes, or attach a PDF…"
-                disabled={isLoading || limitReached}
-                rows={1}
-                aria-label="Message input"
-              />
-              <input
-                ref={fileInputRef}
-                type="file"
-                multiple
-                accept="application/pdf,image/png,image/jpeg,image/gif,image/webp"
-                style={{ display: 'none' }}
-                onChange={(e) => {
-                  if (e.target.files != null && e.target.files.length > 0) {
-                    addFiles(Array.from(e.target.files));
-                  }
-                  e.target.value = '';
-                }}
-                aria-hidden="true"
-                tabIndex={-1}
-              />
-              <button
-                type="button"
-                className={styles.paperclipBtn}
-                aria-label="Attach files"
-                disabled={isLoading || limitReached || chips.length >= MAX_FILE_COUNT}
-                onClick={() => fileInputRef.current?.click()}
-              >
-                <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-                  <path d="m21.44 11.05-9.19 9.19a6 6 0 0 1-8.49-8.49l8.57-8.57A4 4 0 1 1 18 8.84l-8.59 8.57a2 2 0 0 1-2.83-2.83l8.49-8.48" />
-                </svg>
-              </button>
-              <button
-                type="button"
-                className={styles.sendBtn}
-                onClick={() => sendMessage(inputValue)}
-                disabled={!canSend}
-                aria-label="Send message"
-              >
-                <SendIcon width={18} height={18} />
-              </button>
-            </div>
-          </div>
-          {!isPatreon && !limitReached && messagesUsedThisMonth > 0 && (
-            <p className={styles.usageLine}>
-              {remainingMessages === 1
-                ? '1 message left this month — your next send uses it'
-                : `${remainingMessages} of ${FREE_MONTHLY_LIMIT} messages left this month`}
-            </p>
-          )}
-          {networkError != null && (
-            <p className={styles.networkError}>{networkError}</p>
-          )}
-        </div>
+        {loadError != null && (
+          <p style={{ color: 'var(--color-danger)', padding: '0 1.5rem', fontSize: 'var(--text-sm)' }}>
+            {loadError}
+          </p>
+        )}
+        <ChatPanel
+          key={panelSeed.key}
+          initialPrompt={panelSeed.draft}
+          cameFromUpload={cameFromUpload}
+          initialConversationId={panelSeed.conversationId}
+          initialMessages={panelSeed.messages}
+          onConversationCreated={(id, title) => {
+            upsertConversation(id, title);
+            setActiveConversationId(id);
+            setShowEmptyState(false);
+          }}
+          onConversationNotFound={() => {
+            setActiveConversationId(null);
+            setShowEmptyState(true);
+          }}
+        />
       </div>
     </div>
-    </>
   );
 }
